Move module apply_manifest to separate file
Issue-Id: #236
